+++ NEWS.md
@@ -1,5 +1,7 @@
 # vetiver (development version)
 
+* Added support for k-Prototypes clustering from clustMixType (#163, thanks to @galen-ft).
+
 * Now vendor renv directly in package (#157).
 
 * Fixed how `vetiver_ptype()` finds predictors for models (`lm()` and `glm()`) with interactions (#160).